Feature:
Split review section into separate components for better performance.
