+++ src/components/reviews/ExtraContent.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const ExtraContent = () => {
+  return (
+    <div>ExtraContent</div>
+  )
+}
+
+export default ExtraContent
